test verstion
After we generate the graph, we can use funtion GetEdgeCount() to get
the real edge count.
After we read vv graph from disk, we can use funtion GetEdgeCount to get
the real edge count.
When we search, we do not store the real paths,but the paths count. We
can use GetPathsCount() to get paths count.
Add some test code in the FindQualifiedPaths_test funtion to test time
of every part of the funtion.